Refactored more code into lib.rs Mod
