+++ src/main.rs
@@ -1,89 +1,14 @@
 #![cfg_attr(not(debug_assertions), windows_subsystem = "windows")]
 
-use rusqlite::{Connection, Result as SqliteResult};
 use std::env;
 use std::error::Error;
 use std::sync::Arc;
-use machine_info::{Memory, Storage, Processor, Dimension, set_db_path};
+use machine_info::{Memory, Storage, Processor, database};
 slint::include_modules!();
-
-fn init_db() -> SqliteResult<Connection> {
-    let db_path = set_db_path();
-    println!("Using database at: {}", db_path.display());
-    let conn = Connection::open(db_path)?;
-
-    // Create the table only if it doesn't exist
-    conn.execute(
-        "CREATE TABLE IF NOT EXISTS UserSettings (
-            id INTEGER PRIMARY KEY,
-            item_name TEXT NOT NULL,
-            content TEXT NOT NULL,
-            created_at DATETIME DEFAULT CURRENT_TIMESTAMP
-        )",
-        [],
-    )?;
-
-    // Insert default values ONLY if they don't exist (using INSERT OR IGNORE)
-    conn.execute(
-        "INSERT OR IGNORE INTO UserSettings (id, item_name, content)
-         VALUES 
-            (1, 'Default Entry', 'Initial Setting'),
-            (2, 'WindowWidth', '600'),
-            (3, 'WindowHeight', '300')",
-        [],
-    )?;
-
-    Ok(conn)
-}
-
-// Retrieves the currently stored entry from the database
-// Returns: Result containing the stored content string
-fn get_saved_entry(conn: &Connection) -> SqliteResult<String> {
-    conn.query_row("SELECT content FROM UserSettings WHERE id = 1", [], |row| {
-        row.get(0)
-    })
-}
-
-fn set_saved_entry(conn: &Connection, entry: &str) {
-    conn.execute("UPDATE UserSettings SET content = ?1 WHERE id = 1", [entry])
-        .expect("Unable to Save Entry");
-}
-
-fn set_window_position(conn: &Connection, width: i32, height: i32) {
-    conn.execute(
-        "REPLACE INTO UserSettings (id, item_name, content)
-         VALUES
-            (2, 'WindowWidth', ?1),
-            (3, 'WindowHeight', ?2)",
-        [width, height],
-    )
-    .expect("Unable to Save Window Position");
-}
-
-fn get_window_position(conn: &Connection) -> Dimension {
-    let mut my_dimension = Dimension::new();
-    my_dimension.x_position = conn
-        .query_row(
-            "SELECT content FROM UserSettings WHERE item_name = 'WindowWidth'",
-            [],
-            |row| row.get(0),
-        )
-        .expect("Read Failure");
-
-    my_dimension.y_position = conn
-        .query_row(
-            "SELECT content FROM UserSettings WHERE item_name = 'WindowHeight'",
-            [],
-            |row| row.get(0),
-        )
-        .expect("Read Failure");
-    // Return the dimensions
-    my_dimension
-}
 
 fn main() -> Result<(), Box<dyn Error>> {
     // Initialize the database connection and wrap it in Arc (allows multiple conn.executes)
-    let conn = Arc::new(init_db()?);
+    let conn = Arc::new(database::init_db()?);
 
     // Get connection to disks
     let mut _storage_connection = Storage::set_storage_connection();
@@ -101,7 +26,7 @@ fn main() -> Result<(), Box<dyn Error>> {
     let ui = AppWindow::new()?;
 
     // Restore previous session state
-    if let Ok(saved_entry) = get_saved_entry(&conn) {
+    if let Ok(saved_entry) = database::get_saved_entry(&conn) {
         ui.set_input_text(saved_entry.into());
     }
 
@@ -160,7 +85,7 @@ fn main() -> Result<(), Box<dyn Error>> {
             let input_text = ui.get_input_text().to_string();
             #[cfg(debug_assertions)]
             println!("Saving input: {}", input_text);
-            set_saved_entry(&conn, &input_text);
+            database::set_saved_entry(&conn, &input_text);
         }
     });
 
@@ -182,7 +107,7 @@ fn main() -> Result<(), Box<dyn Error>> {
         move || {
             #[cfg(debug_assertions)]
             println!("{:?}", ui_handle.unwrap().window().position());
-            set_window_position(
+            database::set_window_position(
                 &conn,
                 ui_handle.unwrap().window().position().x,
                 ui_handle.unwrap().window().position().y,
@@ -193,7 +118,7 @@ fn main() -> Result<(), Box<dyn Error>> {
 
     // Configure launching of application
     let weak_app = ui.as_weak();
-    let dimensions = get_window_position(&conn);
+    let dimensions = database::get_window_position(&conn);
     slint::invoke_from_event_loop(move || {
         // Set the window position to specific x, y coordinates
         weak_app
